Optimize docs.caseware.com build performance
After some analysis, it looks like changing file permissions is not
necessary.  All that is needed is for the `nginx` user to be able to
access `/usr/share/nginx/docs`.  The files get uploaded as the `jenkins`
user.

We have now added `nginx` to the `jenkins` group on 10.0.0.4.  This
commit updates the `build.sh` script to keep the ownership of all files
as the `jenkins` user.